#22 Редактирование сотрудников. Роли (owner)

+++ resources/views/employees/index.blade.php
@@ -13,11 +13,11 @@
             <div class="col-sm-12">
                 <div class="white-box">
                     <h3 class="box-title pull-left">Сотрудники</h3>
-                    {{--@can('create-'.str_slug('Users'))--}}
+                    @can('create-users')
                         <a class="btn btn-success pull-right" href="{{ route('employees.create') }}">
-                            <i class="icon-plus"></i> Add User
+                            <i class="icon-plus"></i> Новый сотрудник
                         </a>
-                    {{--@endcan--}}
+                    @endcan
                     <div class="clearfix"></div>
                     <hr>
                     <div class="table-responsive">
@@ -34,9 +34,17 @@
                             <tbody>
                             @foreach($users as $user)
                                 <tr>
-                                    <td>{{ $loop->iteration }}</td>
-                                    <td>{{ $user->name }}</td>
-                                    <td>{{ $user->email }}</td>
+                                    <td>{{ $user->id }}</td>
+                                    <td>
+                                        <a href="{{ route('employees.show', $user) }}">
+                                            {{ $user->name }}
+                                        </a>
+                                    </td>
+                                    <td>
+                                        <a href="{{ route('employees.show', $user) }}">
+                                            {{ $user->email }}
+                                        </a>
+                                    </td>
                                     <td>
                                         @foreach($user->roles as $role)
                                             <span class="label label-success">
@@ -45,44 +53,39 @@
                                         @endforeach
                                     </td>
                                     <td>
-                                        {{--@can('show-'.str_slug('Users'))--}}
-                                            <a href="{{ route('employees.show', $user) }}"
-                                               title="View User">
-                                                <button class="btn btn-info btn-sm">
-                                                    <i class="fa fa-eye" aria-hidden="true"></i> View
-                                                </button>
-                                            </a>
-                                        {{--@endcan--}}
+                                        @can('update-users')
+                                            @if(!$user->hasRole('owner'))
+                                                <a href="{{ route('employees.edit', $user) }}"
+                                                   title="Edit User">
+                                                    <button class="btn btn-primary btn-sm">
+                                                        <i class="fa fa-pencil-square-o" aria-hidden="true"> </i>
+                                                    </button>
+                                                </a>
+                                            @endif
+                                        @endcan
 
-                                        {{--@can('update-'.str_slug('Users'))--}}
-                                            <a href="{{ route('employees.edit', $user) }}"
-                                               title="Edit User">
-                                                <button class="btn btn-primary btn-sm">
-                                                    <i class="fa fa-pencil-square-o" aria-hidden="true"> </i> Edit
-                                                </button>
-                                            </a>
-                                        {{--@endcan--}}
-
-                                        {{--@can('delete-'.str_slug('Users'))--}}
-                                            {!! Form::open([
-                                                   'method'=>'DELETE',
-                                                   'route' => ['employees.destroy', $user],
-                                                   'style' => 'display:inline'
-                                               ]) !!}
-                                            {!! Form::button('<i class="fa fa-trash-o" aria-hidden="true"></i> Delete', array(
-                                                    'type' => 'submit',
-                                                    'class' => 'btn btn-danger btn-sm',
-                                                    'title' => 'Delete User',
-                                                    'onclick'=>'return confirm("Confirm delete?")'
-                                            )) !!}
-                                        {{--@endcan--}}
+                                        @can('delete-users')
+                                            @if(!$user->hasRole('owner'))
+                                                {!! Form::open([
+                                                       'method'=>'DELETE',
+                                                       'route' => ['employees.destroy', $user],
+                                                       'style' => 'display:inline'
+                                                   ]) !!}
+                                                {!! Form::button('<i class="fa fa-trash-o" aria-hidden="true"></i>', array(
+                                                        'type' => 'submit',
+                                                        'class' => 'btn btn-danger btn-sm',
+                                                        'title' => 'Delete User',
+                                                        'onclick'=>'return confirm("Confirm delete?")'
+                                                )) !!}
+                                            @endif
+                                        @endcan
                                         {!! Form::close() !!}
                                     </td>
                                 </tr>
                             @endforeach
                             </tbody>
                         </table>
-                        <div class="pagination-wrapper"> {!! $users->appends(['search' => request()->get('search')])->render() !!} </div>
+                        {{--<div class="pagination-wrapper"> {!! $users->appends(['search' => request()->get('search')])->render() !!} </div>--}}
                     </div>
 
                 </div>
